Added cache and removed skeleton + small refactoring
