+++ apps/web/src/app/[locale]/challenge/[id]/submissions/[[...catchAll]]/_components/nosubmissions.tsx
