+++ apps/web/src/app/[locale]/challenge/[id]/submissions/[[...catchAll]]/_components/overview.tsx
