+++ apps/web/src/app/[locale]/challenge/[id]/submissions/[[...catchAll]]/_components/submissions-skeleton.tsx
