+++ apps/web/src/app/[locale]/challenge/[id]/submissions/[[...catchAll]]/getChallengeSubmissions.ts
