--- apps/web/src/app/[locale]/challenge/[id]/submissions/[submissionId]/loading.tsx
@@ -1,1 +1,0 @@
-export { SubmissionSkeleton as default } from './_components/submission-skeleton';
