--- apps/web/src/app/[locale]/challenge/[id]/submissions/loading.tsx
@@ -1,1 +1,0 @@
-export { SubmissionsSkeleton as default } from './_components/submissions-skeleton';
