+++ apps/web/src/app/[locale]/challenge/[id]/wrapper.tsx
@@ -3,8 +3,8 @@ import { useSelectedLayoutSegments } from 'next/navigation';
 import { CodePanel } from '@repo/monaco';
 import { useSession } from '@repo/auth/react';
 import type { ChallengeRouteData } from './getChallengeRouteData';
-import { SubmissionOverview } from './submissions/_components/overview';
-import { saveSubmission } from './submissions/save-submission.action';
+import { SubmissionOverview } from './submissions/[[...catchAll]]/_components/overview';
+import { saveSubmission } from './submissions/[[...catchAll]]/save-submission.action';
 import { FullscreenButton } from '../_components/fullscreen';
 import { ResetEditorButton } from '../_components/reset-editor-button';
 import { EditorShortcutsButton } from '../_components/editor-shortcuts/editor-shortcuts-button';
